[Divider][Material You] convert .js files to .ts
