--- packages/mui-material-next/src/Divider/index.d.ts
@@ -1,5 +1,0 @@
-export { default } from './Divider';
-export * from './Divider';
-
-export { default as dividerClasses } from './dividerClasses';
-export * from './dividerClasses';
