+++ packages/mui-material-next/src/Slider/Slider.tsx
@@ -95,6 +95,7 @@ export const SliderRoot = styled('span', {
     },
   };
 });
+
 export const SliderRail = styled('span', {
   name: 'MuiSlider',
   slot: 'Rail',
