+++ packages/mui-system/src/useThemeProps/getThemeProps.js
@@ -3,12 +3,7 @@ import { internal_resolveProps as resolveProps } from '@mui/utils';
 export default function getThemeProps(params) {
   const { theme, name, props } = params;
 
-  if (
-    !theme ||
-    !theme.components ||
-    !theme.components[name] ||
-    !theme.components[name].defaultProps
-  ) {
+  if (!theme?.components?.[name]?.defaultProps) {
     return props;
   }
